Http server, serving files and dirs.
The server is mostly functional but still has some issues
with waiting for requests sent by the client. It needs
debugging to find out if the browser is sending incompatible
requests which are messing with "std.http" or if our code is
doing something wrong with finishing response headers.

Its possible I am missing something obvious, so it might be a simple
fix.