Refactoring part 2. Generic Locations.
Put a bit of abstraction around the resource locations to accommodate
for resources that may be available through different means other than just simply
being files on the file system. It will improve the flexibility of the
code in the feature if we want to implement things such as HTTP proxies,
forwarding for other protocols such as FTP or SSH or running external
executable.